requirements.txt
# Web Framework
flask==3.0.0
flask[async]==3.0.0
Flask-Limiter==3.5.0
Werkzeug>=3.0.1         # Flask dependency for security features

# AI and Language Processing
openai>=1.10.0,<2.0.0
tiktoken>=0.5.0         # For token counting
langgraph==0.0.21
langchain-core==0.1.17
langchain-openai==0.0.5

# Vector Database & RAG
chromadb>=0.4.0         # Vector database
sentence-transformers>=2.2.0  # Embeddings
faiss-cpu>=1.7.0        # Facebook AI Similarity Search
numpy>=1.24.0           # For vector operations
scikit-learn>=1.3.0     # For additional ML operations

# Performance & Async Processing
aiohttp>=3.9.1         # For async HTTP requests
asyncio-throttle>=1.0.2  # Rate limiting for async calls
httpx>=0.25.0          # Modern async HTTP client
uvloop>=0.19.0         # High-performance event loop (Unix only)

google-generativeai>=0.3.0

# API Frameworks
fastapi>=0.110.0
uvicorn>=0.29.0         # ASGI server for FastAPI

# Document Processing
python-docx==1.1.0
markdown==3.5.2
markdown2==2.4.11
beautifulsoup4==4.12.3
lxml>=4.9.3             # Required by beautifulsoup4 for XML processing
PyPDF2>=3.0.1           # For PDF processing

# Data Validation & Schema
marshmallow==3.20.1
email-validator==2.1.0   # For email validation in forms

# Security
PyJWT>=2.8.0           # For JWT token handling
bcrypt>=4.1.0          # For password hashing

# Storage & Caching
redis>=4.5.1
cachetools>=5.3.0      # In-memory caching
diskcache>=5.6.3       # Persistent disk caching

# Utilities
python-dotenv==1.0.0
colorama==0.4.6
tenacity>=8.2.2        # For retry logic
concurrent-futures>=3.1.1  # For parallel processing
multiprocessing-logging>=0.3.4  # Thread-safe logging

# Web UI & Visualization
streamlit>=1.32.0

# Jira Integration
jira>=3.5.0            # For Jira API integration

# Testing & Development (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black>=23.12.0         # Code formatting
flake8>=7.0.0          # Code linting